@@ -1,11 +1,10 @@
+import React, { Component, useEffect, useRef } from "react";
 
-import React, { Component, useEffect, useRef } from 'react';
-import fs from 'flatstore';
-
-import AlertPanel from './alertpanel';
-import Cell from './cell';
-import PlayerList from './playerlist';
-import Line from './line';
+import AlertPanel from "./alertpanel";
+import Cell from "./cell";
+import PlayerList from "./playerlist";
+import Line from "./line";
+import { bucket } from "react-bucketjs";
 
 // import book1 from '../images/book1.jpg';
 // import book2 from '../images/book2.jpg';
@@ -15,59 +14,68 @@ import Line from './line';
 // import book6 from '../images/book6.jpg';
 // import book7 from '../images/book7.jpg';
 // import test from '../images/test.mp3';
+export const btGameArea = bucket({});
 
 function Gamescreen(props) {
-
     let ref = useRef();
 
     const updatePosition = () => {
-        if (!ref)
-            return;
+        if (!ref) return;
 
         let rect = JSON.stringify(ref.current.getBoundingClientRect());
         rect = JSON.parse(rect);
         rect.offsetWidth = ref.current.offsetWidth;
         rect.offsetHeight = ref.current.offsetHeight;
 
-        fs.set('gamearea', rect);
-    }
+        btGameArea.set(rect);
+    };
 
     useEffect(() => {
         setTimeout(updatePosition, 2000);
-    })
+    });
 
     return (
         <div className="gamewrapper" ref={ref}>
             <div className="vstack">
-                <div className="vstack-noh" >
+                <div className="vstack-noh">
                     <div className="vstack">
                         <PlayerList />
-
                     </div>
 
                     <AlertPanel />
                 </div>
-                <div className="gamescreen" >
-                    <Line className={'foreground'} />
+                <div className="gamescreen">
+                    <Line className={"foreground"} />
                     {/* <Line className={'foreground'} />
                         <Line className={'background'} />
                         <Line className={'background'} /> */}
-                    <Line className={'background'} />
+                    <Line className={"background"} />
                     <div className="gamearea">
                         <div className="vstack">
                             <div className="hstack">
-                                <Cell id={0} /><div className="vertical" /><Cell id={1} /><div className="vertical" /><Cell id={2} />
+                                <Cell id={0} />
+                                <div className="vertical" />
+                                <Cell id={1} />
+                                <div className="vertical" />
+                                <Cell id={2} />
                             </div>
                             <div className="horizontal" />
                             <div className="hstack">
-                                <Cell id={3} /><div className="vertical" /><Cell id={4} /><div className="vertical" /><Cell id={5} />
+                                <Cell id={3} />
+                                <div className="vertical" />
+                                <Cell id={4} />
+                                <div className="vertical" />
+                                <Cell id={5} />
                             </div>
                             <div className="horizontal" />
                             <div className="hstack">
-                                <Cell id={6} /><div className="vertical" /><Cell id={7} /><div className="vertical" /><Cell id={8} />
+                                <Cell id={6} />
+                                <div className="vertical" />
+                                <Cell id={7} />
+                                <div className="vertical" />
+                                <Cell id={8} />
                             </div>
                         </div>
-
                     </div>
 
                     {/* <audio controls >
@@ -80,14 +88,10 @@ function Gamescreen(props) {
                 <img src={book5} />
                 <img src={book6} />
                 <img src={book7} /> */}
-
-
                 </div>
             </div>
         </div>
-    )
-
-
+    );
 }
 
 export default Gamescreen;
