@@ -1,4 +1,6 @@
-import cup from './acosg';
+// import ACOSServer from './acosg';
+
+import { ACOSServer } from "acosgames";
 
 // let defaultGame = {
 //     state: {
@@ -10,25 +12,24 @@ import cup from './acosg';
 //     },
 //     players: {},
 //     next: {},
-//     events: {} 
+//     events: {}
 // }
 
-let defaultState = { cells: ['', '', '', '', '', '', '', '', ''], }
+// let defaultState = { cells: ["", "", "", "", "", "", "", "", ""] };
 
 class Tictactoe {
+    defaultState = { cells: ["", "", "", "", "", "", "", "", ""] };
 
     onNewGame(action) {
-
-        let game = cup.game();
-        game.state = defaultState;
+        let game = ACOSServer.gamestate();
+        game.state = this.defaultState;
 
         this.checkNewRound();
     }
 
     onSkip(action) {
-        let next = cup.next();
-        if (!next || !next.id)
-            return;
+        let next = ACOSServer.next();
+        if (!next || !next.id) return;
         // let id = action.payload.id;
         // if (!next.id) {
         //     id = next.id;
@@ -38,48 +39,41 @@ class Tictactoe {
     }
 
     onJoin(action) {
-        cup.log(action);
-        if (!action.user.id)
-            return;
+        ACOSServer.output(action);
+        if (!action.user.id) return;
 
-        let player = cup.players(action.user.id);
+        let player = ACOSServer.players(action.user.id);
         // player.rank = 2;
         // player.score = 0;
 
-        let teams = cup.teams();
+        let teams = ACOSServer.teams();
         if (teams && player.teamid) {
             teams[player.teamid].rank = 2;
             teams[player.teamid].score = 0;
         }
 
+        ACOSServer.output("TICTACTOE PLAYER INFO:");
+        ACOSServer.output(player);
 
-        cup.log("TICTACTOE PLAYER INFO:");
-        cup.log(player);
+        ACOSServer.players(action.user.id, player);
 
-        cup.players(action.user.id, player);
-
-        let playerCount = cup.playerCount();
+        let playerCount = ACOSServer.playerCount();
         if (playerCount <= 2) {
-            cup.event('join', {
-                id: action.user.id
+            ACOSServer.event("join", {
+                id: action.user.id,
             });
             // this.checkNewRound();
-        }
-        else {
-
+        } else {
         }
 
-
-        // if (cup.players(action.user.id).type)
+        // if (ACOSServer.players(action.user.id).type)
         //     return;
-
-
     }
 
     checkNewRound() {
         //if player count reached required limit, start the game
-        //let maxPlayers = cup.rules('maxPlayers') || 2;
-        let playerCount = cup.playerCount();
+        //let maxPlayers = ACOSServer.rules('maxPlayers') || 2;
+        let playerCount = ACOSServer.playerCount();
         if (playerCount >= 2) {
             this.newRound();
         }
@@ -90,7 +84,7 @@ class Tictactoe {
     }
 
     playerLeave(id) {
-        let players = cup.players();
+        let players = ACOSServer.players();
         let otherPlayerId = null;
         if (players[id]) {
             otherPlayerId = this.selectNextPlayer(id);
@@ -99,33 +93,30 @@ class Tictactoe {
 
         if (otherPlayerId) {
             let otherPlayer = players[otherPlayerId];
-            this.setWinner(otherPlayer.type, 'forfeit')
+            this.setWinner(otherPlayer.type, "forfeit");
         }
     }
 
     onPick(action) {
-        let room = cup.room();
-        if (room.status != 'gamestart')
-            return false;
+        let room = ACOSServer.room();
+        if (room.status != "gamestart") return false;
 
-        let state = cup.state();
-        let user = cup.players(action.user.id);
-        if (user.test2)
-            delete user.test2;
+        let state = ACOSServer.state();
+        let user = ACOSServer.players(action.user.id);
+        if (user.test2) delete user.test2;
         //get the picked cell
         let cellid = action.payload;
-        if (typeof cellid !== 'number')
-            return false;
+        if (typeof cellid !== "number") return false;
 
         let cell = state.cells[cellid];
 
         // block picking cells with markings, and send error
         if (cell.length > 0) {
-            cup.next({
+            ACOSServer.next({
                 id: action.user.id,
-                action: 'pick',
-                error: 'NOT_EMPTY'
-            })
+                action: "pick",
+                error: "NOT_EMPTY",
+            });
             return false;
         }
 
@@ -134,64 +125,63 @@ class Tictactoe {
         let id = action.user.id;
         state.cells[cellid] = type;
 
-        // cup.event('picked', {
+        // ACOSServer.event('picked', {
         //     cellid, id
         // });
-        // cup.prev()
+        // ACOSServer.prev()
 
         if (this.checkWinner()) {
             return;
         }
 
-        cup.setTimelimit(15);
+        ACOSServer.setTimelimit(15);
         this.selectNextPlayer(null);
     }
 
     newRound() {
-        let playerList = cup.playerList();
+        let playerList = ACOSServer.playerList();
 
-        let state = cup.state();
+        let state = ACOSServer.state();
         //select the starting player
 
-        let xteam = cup.teams('team_x');
-        // let xplayer = cup.players(  );
+        let xteam = ACOSServer.teams("team_x");
+        // let xplayer = ACOSServer.players(  );
 
-        state.sx = xteam.players[0]
+        state.sx = xteam.players[0];
 
-        cup.next({
+        ACOSServer.next({
             id: state.sx,
-            action: 'pick'
+            action: "pick",
         });
-        let players = cup.players() || {};
+        let players = ACOSServer.players() || {};
         let playerIds = Object.keys(players);
 
-        let otherIds = playerIds.filter(value => value != state.sx);
+        let otherIds = playerIds.filter((value) => value != state.sx);
 
-        // let x = cup.randomInt(0, 2);
+        // let x = ACOSServer.randomInt(0, 2);
         // let o = x == 0 ? 1 : 0;
-        let playerX = state.sx;//playerIds[x];
-        let playerO = otherIds[0];//playerIds[o];
+        let playerX = state.sx; //playerIds[x];
+        let playerO = otherIds[0]; //playerIds[o];
 
-        players[playerX].type = 'X';
-        players[playerO].type = 'O';
+        players[playerX].type = "X";
+        players[playerO].type = "O";
 
-        cup.event('newround', true);
-        cup.setTimelimit(15);
+        ACOSServer.event("newround", true);
+        ACOSServer.setTimelimit(15);
     }
 
     selectNextPlayer(userid) {
-        let action = cup.action();
-        let players = cup.playerList();
+        let action = ACOSServer.action();
+        let players = ACOSServer.playerList();
         userid = userid || action.user.id;
         //only 2 players so just filter the current player
-        let remaining = players.filter(x => x != userid);
-        cup.next({
+        let remaining = players.filter((x) => x != userid);
+        ACOSServer.next({
             id: remaining[0],
-            action: 'pick'
+            action: "pick",
         });
         return remaining[0];
     }
-
 
     // Check each strip that makes a win
     //      0  |  1  |  2
@@ -213,12 +203,12 @@ class Tictactoe {
     }
 
     checkNoneEmpty() {
-        let cells = cup.state().cells;
+        let cells = ACOSServer.state().cells;
         // let cellslist = [];
         // for (var key in cells) {
         //     cellslist.push(cells[key]);
         // }
-        let filtered = cells.filter(v => v == '');
+        let filtered = cells.filter((v) => v == "");
 
         if (filtered.length == 0) {
             this.setTie();
@@ -228,17 +218,15 @@ class Tictactoe {
 
     // checks if a strip has matching types
     check(strip) {
-        let cells = cup.state().cells;
+        let cells = ACOSServer.state().cells;
         // let cellslist = [];
         // for (var key in cells) {
         //     cellslist.push(cells[key]);
         // }
 
-
         let first = cells[strip[0]];
-        if (first == '')
-            return false;
-        let filtered = strip.filter(id => cells[id] == first);
+        if (first == "") return false;
+        let filtered = strip.filter((id) => cells[id] == first);
         if (filtered.length == 3 && filtered.length == strip.length) {
             this.setWinner(first, strip);
             return true;
@@ -247,30 +235,28 @@ class Tictactoe {
     }
 
     findPlayerWithType(type) {
-        let players = cup.players();
+        let players = ACOSServer.players();
         for (var id in players) {
             let player = players[id];
-            if (player.type == type)
-                return id;
+            if (player.type == type) return id;
         }
         return null;
     }
 
-
     setTie() {
-        cup.gameover({ type: 'tie' })
-        cup.next({});
-        // cup.prev({})
+        ACOSServer.gameover({ type: "tie" });
+        ACOSServer.next({});
+        // ACOSServer.prev({})
 
-        // cup.killGame();
+        // ACOSServer.killGame();
     }
     // set the winner event and data
     setWinner(type, strip) {
         //find user who matches the win type
         let userid = this.findPlayerWithType(type);
-        let player = cup.players(userid);
+        let player = ACOSServer.players(userid);
 
-        let teams = cup.teams();
+        let teams = ACOSServer.teams();
         if (teams && player.teamid) {
             teams[player.teamid].rank = 1;
             teams[player.teamid].score = 100;
@@ -278,18 +264,18 @@ class Tictactoe {
         // player.rank = 1;
         // player.score = player.score + 100;
         if (!player) {
-            player.id = 'unknown player';
+            player.id = "unknown player";
         }
 
-        cup.gameover({
-            type: 'winner',
+        ACOSServer.gameover({
+            type: "winner",
             pick: type,
             strip: strip,
-            id: userid
+            id: userid,
         });
         // cup.prev()
-        cup.next({});
-        // cup.killGame();
+        ACOSServer.next({});
+        // ACOSServer.killGame();
     }
 }
 
